requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
python-dotenv>=1.0.0
slack-sdk>=3.20.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
certifi>=2024.12.0
